Extract selection to module
